+++ annostat/__init__.py
@@ -1,3 +1,3 @@
 """Tools for analyzing bacterial genome annotations."""
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"